Задание 3_unit_test
