+++ OmSTU-AMCS-SummerPractice2023.sln
@@ -1,0 +1,22 @@
+
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SquareEquation.Tests", "SquareEquation.Tests\SquareEquation.Tests.csproj", "{97AD715A-E8B7-4BC0-943D-1A42B5F6143F}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{97AD715A-E8B7-4BC0-943D-1A42B5F6143F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{97AD715A-E8B7-4BC0-943D-1A42B5F6143F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{97AD715A-E8B7-4BC0-943D-1A42B5F6143F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{97AD715A-E8B7-4BC0-943D-1A42B5F6143F}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
