--- unit-testing-using-dotnet-test/OmSTU-AMCS-SummerPractice2023.sln
@@ -1,28 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SquareEquationLib", "SquareEquationLib\SquareEquationLib.csproj", "{2E1F98E7-1CEB-443D-90CD-A9C86287E639}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SquareEquationLib.Tests", "SquareEquationLib.Tests\SquareEquationLib.Tests.csproj", "{DB2EDBAB-5E73-46BF-BCDB-C3702536608D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2E1F98E7-1CEB-443D-90CD-A9C86287E639}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2E1F98E7-1CEB-443D-90CD-A9C86287E639}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2E1F98E7-1CEB-443D-90CD-A9C86287E639}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2E1F98E7-1CEB-443D-90CD-A9C86287E639}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DB2EDBAB-5E73-46BF-BCDB-C3702536608D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DB2EDBAB-5E73-46BF-BCDB-C3702536608D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DB2EDBAB-5E73-46BF-BCDB-C3702536608D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DB2EDBAB-5E73-46BF-BCDB-C3702536608D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
