--- unit-testing-using-dotnet-test/unit-testing-using-dotnet-test.sln
@@ -1,28 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.31903.59
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PrimeService", "PrimeService\PrimeService.csproj", "{63FBB1F7-F95A-4330-93E6-CDBC09711DF4}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PrimeService.Tests", "PrimeService.Tests\PrimeService.Tests.csproj", "{30DBF16C-E0F9-4EEA-8FD3-AB4C8E221485}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{63FBB1F7-F95A-4330-93E6-CDBC09711DF4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{63FBB1F7-F95A-4330-93E6-CDBC09711DF4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{63FBB1F7-F95A-4330-93E6-CDBC09711DF4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{63FBB1F7-F95A-4330-93E6-CDBC09711DF4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{30DBF16C-E0F9-4EEA-8FD3-AB4C8E221485}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{30DBF16C-E0F9-4EEA-8FD3-AB4C8E221485}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{30DBF16C-E0F9-4EEA-8FD3-AB4C8E221485}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{30DBF16C-E0F9-4EEA-8FD3-AB4C8E221485}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-EndGlobal
